Update state merkle tree
