+++ state/merkle.go
@@ -4,37 +4,47 @@ import (
 	"github.com/zarbchain/zarb-go/account"
 	"github.com/zarbchain/zarb-go/crypto"
 	simpleMerkle "github.com/zarbchain/zarb-go/libs/merkle"
+	"github.com/zarbchain/zarb-go/logger"
 	"github.com/zarbchain/zarb-go/validator"
 )
 
 func (st *state) accountsMerkleRootHash() crypto.Hash {
-	accs := make([]*account.Account, 0)
+	total := st.store.TotalAccounts()
+	hashes := make([]crypto.Hash, total)
 	st.store.IterateAccounts(func(acc *account.Account) (stop bool) {
-		accs = append(accs, acc)
+		if acc.Number() >= total {
+			panic("Account number is out of range")
+		}
+		hashes[acc.Number()] = acc.Hash()
 		return false
 	})
-
-	hashes := make([]crypto.Hash, len(accs))
-	for i, a := range accs {
-		hashes[i] = a.Hash()
-	}
 
 	tree := simpleMerkle.NewTreeFromHashes(hashes)
 	return tree.Root()
 }
 
 func (st *state) validatorsMerkleRootHash() crypto.Hash {
-	vals := make([]*validator.Validator, 0)
+	total := st.store.TotalValidators()
+	hashes := make([]crypto.Hash, total)
 	st.store.IterateValidators(func(val *validator.Validator) (stop bool) {
-		vals = append(vals, val)
+		if val.Number() >= total {
+			panic("Account number is out of range")
+		}
+		hashes[val.Number()] = val.Hash()
 		return false
 	})
-
-	hashes := make([]crypto.Hash, len(vals))
-	for i, v := range vals {
-		hashes[i] = v.Hash()
-	}
-
 	tree := simpleMerkle.NewTreeFromHashes(hashes)
 	return tree.Root()
+}
+
+func (st *state) stateHash() crypto.Hash {
+	accRootHash := st.accountsMerkleRootHash()
+	valRootHash := st.validatorsMerkleRootHash()
+
+	rootHash := simpleMerkle.HashMerkleBranches(&accRootHash, &valRootHash)
+	if rootHash == nil {
+		logger.Panic("State hash can't be nil")
+	}
+
+	return *rootHash
 }
